term-6502/requirements.txt
pyserial
coloredlogs